Get rid of TransactionService and remap to Cache.
