+++ CacheService/docker_context/cacheInputSchema.py
@@ -45,17 +45,16 @@ funds_schema = {
     'required': ['transaction_num', 'user_id', 'amount', 'command']
 }
 
-stocks_query_schema = {
+base_transaction_schema = {
     'type': 'object',
     'properties': {
         'transaction_num': {'type': 'integer'},
         'user_id': {'type': 'string'},
         'stock_symbol': one_to_three_letter_string,
-        'stock_amount': non_negative_number,
-        'funds': non_negative_number,
+        'amount': non_negative_number,
         'command': {'type': 'string'}
     },
-    'required': ['transaction_num', 'user_id', 'stock_symbol', 'stock_amount', 'funds', 'command']
+    'required': ['transaction_num', 'user_id', 'stock_symbol', 'amount', 'command']
 }
 
 user_schema = {
@@ -85,7 +84,25 @@ update_stock_schema = {
     'required': ['user_id', 'stock_symbol'],
 }
 
+quote_schema = {
+    'type': 'object',
+    'properties': {
+        'transaction_num': {'type': 'integer'},
+        'user_id': {'type': 'string'},
+        'stock_symbol': one_to_three_letter_string,
+        'command': {'type': 'string'}
+    },
+    'required': ['transaction_num', 'user_id', 'stock_symbol', 'command']
+}
+
 add_funds_schema = funds_schema
 remove_funds_schema = funds_schema
-buy_stock_schema = stocks_query_schema
-sell_stock_schema = stocks_query_schema
+
+buy_stock_schema = base_transaction_schema
+commit_buy_schema = user_schema
+cancel_buy_schema = user_schema
+
+sell_stock_schema = base_transaction_schema
+commit_sell_schema = user_schema
+cancel_sell_schema = user_schema
+
