--- TransactionService/docker_context/requirements.txt
@@ -1,5 +1,0 @@
-asyncio==3.4.3
-sanic==20.12.1
-aiohttp==3.7.3
-jsonschema==3.2.0
-aioredis==1.3.0
