--- TransactionService/stop_service.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-
-source $(dirname $0)/config.sh
-
-cont=$trans_service_name
-
-echo 'Stopping transaction service...'
-sudo docker stop $cont > /dev/null
