+++ build_service.sh
@@ -8,5 +8,4 @@ $curr_path/DatabaseService/build_service.sh 2> /dev/null
 $curr_path/AuditService/build_service.sh 2> /dev/null
 $curr_path/TriggerService/build_service.sh 2> /dev/null
 $curr_path/CacheService/build_service.sh 2> /dev/null
-$curr_path/TransactionService/build_service.sh 2> /dev/null
 $curr_path/WebService/build_service.sh 2> /dev/null
