+++ start_service.sh
@@ -13,5 +13,4 @@ $curr_path/DatabaseService/start_service.sh
 $curr_path/AuditService/start_service.sh
 $curr_path/TriggerService/start_service.sh
 $curr_path/CacheService/start_service.sh
-$curr_path/TransactionService/start_service.sh
 $curr_path/WebService/start_service.sh
