+++ stop_service.sh
@@ -3,7 +3,6 @@
 curr_path=$(dirname $0)
 
 $curr_path/WebService/stop_service.sh 2> /dev/null
-$curr_path/TransactionService/stop_service.sh 2> /dev/null
 $curr_path/CacheService/stop_service.sh 2> /dev/null
 $curr_path/TriggerService/stop_service.sh 2> /dev/null
 $curr_path/DummyStockServer/stop_service.sh 2> /dev/null
